Chris push
made player jump, fixed camera angle (still to be adjusted), fixed camera speed, fixed collision on front of the castle, being able to enter through it now. Added canvas, currently a plain text of score, and a button that is linked to the repository. Added a script for health, if player health goes to 0, the object is destroyed.

Next update road :
Player model
Player health and score functionality
Player health and score display functionality
Add pickup prefabs and collision scripts for them
Improve the overall look of the game, add houses, objects, miscellanous objects, improve terrain.
Player needs a model and animations
Create an enemy, add models and animations, health and ability to damage the player either by collision (either running at them or shooting prefabs, if prefabs collide with player, reduce health)
Add a timer for the level for the player to feel pressure on the game to finish
Create an object to spawn after a certain score reached by player for the next level